added heartbeat
main code is in general_ui, subtracting score from an arbitrary number as delay for a sound loop, then once reaching a threshold, switching to static fast loop. can change the '6' in both the subtraction/greater= to change the delay, lower to have heartbeat get faster quicker, or higher to slow that process down.